Refactor: Extract TaskNode and AutoReleaseContinuationBox to separate files
- Extract TaskNode class from TaskQueueActor.swift to TaskNode.swift
- Extract AutoReleaseContinuationBox class from TaskQueueActor.swift to AutoReleaseContinuationBox.swift
- Improve code organization and maintainability
